+++ include/seqan3/search/fm_index/bi_fm_index.hpp
@@ -17,7 +17,6 @@
 #include <utility>
 
 #include <seqan3/core/range/type_traits.hpp>
-#include <seqan3/range/views/persist.hpp>
 #include <seqan3/search/fm_index/fm_index.hpp>
 #include <seqan3/search/fm_index/bi_fm_index_cursor.hpp>