+++ test/unit/utility/views/repeat_test.cpp
@@ -11,8 +11,8 @@
 #include <seqan3/std/ranges>
 
 #include <seqan3/core/detail/iterator_traits.hpp>
+#include <seqan3/core/detail/persist_view.hpp>
 #include <seqan3/core/range/type_traits.hpp>
-#include <seqan3/range/views/persist.hpp>
 #include <seqan3/range/views/take.hpp>
 #include <seqan3/range/views/take_exactly.hpp>
 #include <seqan3/test/expect_range_eq.hpp>
@@ -156,7 +156,7 @@ TEST(view, factory)
 
     // view
     {
-        auto view = std::string{"foobar"} | seqan3::views::persist | std::views::take(3);
+        auto view = std::string{"foobar"} | seqan3::detail::persist | std::views::take(3);
         auto v = seqan3::views::repeat(view);
         EXPECT_RANGE_EQ(*v.begin(), view);
     }