+++ CHANGELOG.md
@@ -200,7 +200,8 @@ regression test suite and patches at https://github.com/seqan/seqan3/tree/master
 
 * We made `seqan3::views::convert` NOAPI and moved it to `seqan3/utility/views/convert.hpp`. You can still use
   `seqan3::views::convert` in the meantime, but we encourage using `std::views::transform` instead as shown in our
-  [Cookbook](https://docs.seqan.de/seqan/3.0.3/cookbook.html#cookbook_convert_alphabet_range).
+  [Cookbook](https://docs.seqan.de/seqan/3.0.3/cookbook.html#cookbook_convert_alphabet_range)
+  ([\#2524](https://github.com/seqan/seqan3/pull/2524)).
 * Deprecated `seqan3::views::drop`, use `std::views::drop` or `seqan3::views::type_reduce | std::views::drop`.
   ([\#2540](https://github.com/seqan/seqan3/pull/2540))
 * We deprecated `seqan3::views::to_upper` and it will be removed in 3.1.0, use
@@ -209,6 +210,8 @@ regression test suite and patches at https://github.com/seqan/seqan3/tree/master
 * We deprecated `seqan3::views::to_lower` and it will be removed in 3.1.0, use
    `std::views::transform([](auto && chr){return std::tolower(chr)})`.
    ([\#2556](https://github.com/seqan/seqan3/pull/2556))
+* Deprecated `seqan3::views::persist`. There is no replacement, use lvalues instead of rvalues
+  ([\#2553](https://github.com/seqan/seqan3/pull/2553)).
 
 #### Search